Misc updates on DTOs

+++ src/main/java/com/example/xpmovies/dto/MovieCreateDto.java
@@ -1,5 +1,6 @@
 package com.example.xpmovies.dto;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import lombok.Data;
@@ -11,7 +12,7 @@ public class MovieCreateDto {
 
 	private String title;
 	private Date launchDate;
-	private int rank;
-	private Double revenue;
+	private Double rank;
+	private BigDecimal revenue;
 	
 }
